# contracts/document/src/models/CreateSectionDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateSectionDto = {
    /**
     * Display name of the section
     */
    sectionTitle: string;
};