# contracts/document/src/models/CreateWorkshopPageDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateWorkshopPageDto = {
    workshopId: string;
    name?: string;
    sortId?: number;
    pageType?: string;
    lastUpdated?: string;
    /**
     * Serialized JSON of the document blocks
     */
    html?: string;
};