# contracts/document/src/models/DeletePageParamsDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DeletePageParamsDto = {
    _id: string;
    workshopId: string;
    name: string;
};