# contracts/document/src/models/EditPageNameUpdateWorkshopDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type EditPageNameUpdateWorkshopDto = {
    _id: string;
    name: string;
    workshopId: string;
};