# contracts/document/src/models/WorkshopPageDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type WorkshopPageDto = {
    _id: string;
    workshopGroupId: string;
    name: string;
    sortId: number;
    pageType: string;
    lastUpdated: string;
    /**
     * Serialized JSON of the document blocks
     */
    html: string;
    __v: number;
};