# contracts/document/src/models/WorkshopPageIdentifierDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type WorkshopPageIdentifierDto = {
    /**
     * Object id of the workshop-page.
     */
    _id: string;
    /**
     * Display name of the workshop-page.
     */
    name: string;
    /**
     * Position of the doc item in the section's list
     */
    sortId: number;
};